requirements.txt
appnope==0.1.0
astroid==1.6.3
backcall==0.1.0
bleach==2.1.3
certifi==2018.1.18
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
Flask==1.0.1
Flask-Cors==3.0.4
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.0
isort==4.3.4
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
louvain==0.6.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.1
numpy==1.14.2
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycairo==1.17.0
Pygments==2.2.0
pylint==1.8.4
pyparsing==2.2.0
python-dateutil==2.7.2
python-igraph==0.7.1.post6
pytz==2018.4
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scipy==1.0.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.0
wrapt==1.10.11
